Store data to Oracle

+++ src/API/LCR.Import.Web.Api/Resources/Commands/ProccessFileCommand.cs
@@ -5,8 +5,8 @@ namespace LCR.Import.Web.Api.Resources
   public class ProccessFileCommand : ICommand
   {
     public Guid CorrelationId { get; set; }
-    public int UserId { get; set; }
-    public int ImportHistoryId { get; set; }
+    public decimal UserId { get; set; }
+    public decimal ImportHistoryId { get; set; }
     public string FilePath { get; set; }
   }
 }
