+++ src/BuildingBlocks/DataAccess/LCR.TPM.Model/Import/ImportMappedDataModel.cs
@@ -6,9 +6,9 @@ namespace LCR.TPM.Model
   public class ImportMappedDataModel : BaseModel
   {
     public UploadHistoryModel UploadHistory { get; set; }
-    public int UploadHistoryId { get; set; }
+    public decimal UploadHistoryId { get; set; }
     public ImportRawDataModel ImportRawData { get; set; }
-    public int ImportRawDataId { get; set; }
+    public decimal ImportRawDataId { get; set; }
     public int FileOperatorId { get; set; }
     public char FileDirection { get; set; }
     public DateTime FileDateOpen { get; set; }
